+++ modules/main/src/main/java/com/nisovin/shopkeepers/config/Settings.java
@@ -93,6 +93,8 @@ public class Settings extends Config {
 
 	public static boolean enableTownyRestrictions = false;
 
+	public static boolean enableItemsAdderIntegration = true;
+
 	public static boolean checkSpawnLocationInteractionResult = false;
 
 	public static boolean disableInventoryVerification = false;